backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
pydantic[email]
psycopg2-binary
python-multipart
pillow
rembg
onnxruntime
colorthief
alembic
torch
transformers
Pillow
bcrypt
python-jose[cryptography]
